+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Transitions between screens.
+- Screen shake effect.
 
 ## [1.1.5] - 2024-02-08